% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_oncotree_versions}
\alias{get_oncotree_versions}
\title{Get OncoTree Versions}
\usage{
get_oncotree_versions()
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Get OncoTree Versions
}
\seealso{
Other get functions: 
\code{\link{get_main_types}()},
\code{\link{get_tumor_types_tree}()},
\code{\link{get_tumor_types_txt}()},
\code{\link{get_tumor_types}()}
}
\concept{get functions}
