% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{get_tumor_types_txt}
\alias{get_tumor_types_txt}
\title{Get tumorTypes Text}
\usage{
get_tumor_types_txt(version = "oncotree_latest_stable")
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Return tumor types in plain text format
}
\seealso{
Other get functions: 
\code{\link{get_main_types}()},
\code{\link{get_oncotree_versions}()},
\code{\link{get_tumor_types_tree}()},
\code{\link{get_tumor_types}()}
}
\concept{get functions}
