% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup_oncotree_code_locally.R
\name{lookup_oncotree_code_locally}
\alias{lookup_oncotree_code_locally}
\title{Lookup an OncoTree Code}
\usage{
lookup_oncotree_code_locally(oncotree_code)
}
\arguments{
\item{oncotree_code}{vector of length 1}
}
\description{
Lookup an OncoTree Code
}
