% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lookup.R
\name{lookup_umls}
\alias{lookup_umls}
\title{Lookup UMLS CUI}
\usage{
lookup_umls(umls_cui, exact_match = TRUE, level = 1:5)
}
\arguments{
\item{umls_cui}{UMLS CUI from the UMLS Metathesaurus to look up.}

\item{exact_match}{If it sets to true, it will only return one element array.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Lookup UMLS CUI
}
\seealso{
Other lookup functions: 
\code{\link{lookup_nci}()}
}
\concept{lookup functions}
