% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/args.R
\name{pkg_args}
\alias{pkg_args}
\title{Package Arguments}
\arguments{
\item{version}{OncoTree version. One of "oncotree_latest_stable," "oncotree_candidate_release," or "oncotree_development".}

\item{type}{one of "code", "name", "mainType", "level", "nci", "umls", or "color"}

\item{search}{character vector of length one of the query}

\item{exact_match}{If it sets to true, it will only return one element array.}

\item{levels}{vector of integers ranging from 1:5. Level 1 is equivalent to the Tissue Level.}
}
\description{
Package Arguments
}
