% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_level}
\alias{search_level}
\title{Search level}
\usage{
search_level(level)
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Return OncoTree by Level
}
\seealso{
Other search functions: 
\code{\link{search_main_type}()},
\code{\link{search_name}()}
}
\concept{search functions}
