% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_main_type}
\alias{search_main_type}
\title{Search mainType}
\usage{
search_main_type(search, exact_match = TRUE, level = 1:5)
}
\arguments{
\item{search}{character vector of length one of the query}

\item{exact_match}{If it sets to true, it will only return one element array.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Search mainType. To return all possible mainTypes, see \code{\link{get_main_types}}.
}
\seealso{
Other search functions: 
\code{\link{search_level}()},
\code{\link{search_name}()}
}
\concept{search functions}
