% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_name}
\alias{search_name}
\title{Search Name}
\usage{
search_name(search, exact_match = TRUE, level = 1:5)
}
\arguments{
\item{search}{character vector of length one of the query}

\item{exact_match}{If it sets to true, it will only return one element array.}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Search Name
}
\seealso{
Other search functions: 
\code{\link{search_level}()},
\code{\link{search_main_type}()}
}
\concept{search functions}
