% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_nci.R
\name{search_nci}
\alias{search_nci}
\title{Search NCI}
\usage{
search_nci(search, exact_match = TRUE, levels = 1:5)
}
\arguments{
\item{search}{character vector of length one of the query}

\item{exact_match}{If it sets to true, it will only return one element array.}

\item{levels}{vector of integers ranging from 1:5}
}
\description{
Search NCI
}
