% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_tumor_types}
\alias{search_tumor_types}
\title{Search Tumor Types}
\usage{
search_tumor_types(
  type,
  search,
  exact_match = TRUE,
  version = "oncotree_latest_stable",
  level = 1:5
)
}
\arguments{
\item{type}{one of "code", "name", "mainType", "level", "nci", "umls", or "color"}

\item{search}{character vector of length one of the query}

\item{exact_match}{If it sets to true, it will only return one element array.}

\item{version}{OncoTree version. One of "oncotree_latest_stable," "oncotree_candidate_release," or "oncotree_development".}
}
\value{
a [tibble][tibble::tibble-package]
}
\description{
Primitive for all `search_*` and `lookup_*` functions in this package.
}
\seealso{
\code{\link[urltools]{url_decode}}
 \code{\link[httr]{GET}},\code{\link[httr]{content}},\code{\link[httr]{http_error}}
 \code{\link[jsonlite]{toJSON, fromJSON}}
}
