% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_umls.R
\name{search_umls}
\alias{search_umls}
\title{Search UMLS}
\usage{
search_umls(search, exact_match = TRUE, levels = 1:5)
}
\arguments{
\item{search}{character vector of length one of the query}

\item{exact_match}{If it sets to true, it will only return one element array.}

\item{levels}{vector of integers ranging from 1:5}
}
\description{
Search UMLS
}
