% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_level.R
\name{search_level}
\alias{search_level}
\title{Search level}
\usage{
search_level(level, exact_match = TRUE)
}
\arguments{
\item{level}{vector of length 1}

\item{exact_match}{If it sets to true, it will only return one element array.}
}
\description{
Search level
}
