% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_tumor_types.R
\name{search_tumor_types}
\alias{search_tumor_types}
\title{Search Tumor Types}
\usage{
search_tumor_types(type, search, exact_match = TRUE, levels = 1:5)
}
\arguments{
\item{type}{one of "code", "name", "mainType", "level", "nci", "umls", or "color"}

\item{search}{character vector of length one of the query}

\item{exact_match}{If it sets to true, it will only return one element array.}

\item{levels}{vector of integers ranging from 1:5}
}
\description{
Search Tumor Types
}
